pyproject.toml
[project]
name = "mlp"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "torch",
    "transformers",
    "faiss-gpu-cu12",
    "tqdm",
    "numpy",
    "accelerate",
    "safetensors",
    "pdfminer-six>=20260107",
    "requests>=2.32.5",
    "python-dotenv>=1.2.2",
    "google-genai>=1.0.0",
]
